refactor(ui): Break phone number utils and data into proper folders
